requirements.txt
astroid==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.9
cobble==0.1.3
colorama==0.4.4
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mammoth==1.4.18
mccabe==0.6.1
platformdirs==2.4.1
ply==3.11
pylint==2.12.2
requests==2.26.0
six==1.16.0
smmap==5.0.0
stone==3.2.1
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
wrapt==1.13.3
